Move length calculation to a separate class
