+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/util/SimpleOffsetSerializer.java
@@ -80,8 +80,8 @@ import tech.pegasys.teku.ssz.sos.SimpleOffsetSerializable;
 @SuppressWarnings({"rawtypes", "unchecked"})
 public class SimpleOffsetSerializer {
 
-  private static final int UNSIGNED_LONG_SIZE = 8;
-  private static final int BOOLEAN_SIZE = 1;
+  static final int UNSIGNED_LONG_SIZE = 8;
+  static final int BOOLEAN_SIZE = 1;
   public static HashMap<Class, ReflectionInformation> classReflectionInfo = new HashMap<>();
   public static HashMap<Class, LengthBounds> classLengthBounds = new HashMap<>();
 
@@ -127,7 +127,7 @@ public class SimpleOffsetSerializer {
     }
 
     for (Class classItem : classes) {
-      classLengthBounds.put(classItem, calculateLengthBounds(classItem));
+      classLengthBounds.put(classItem, LengthBoundCalculator.calculateLengthBounds(classItem));
     }
   }
 
@@ -230,112 +230,6 @@ public class SimpleOffsetSerializer {
 
   public static <T> LengthBounds getLengthBounds(final Class<T> type) {
     return checkNotNull(classLengthBounds.get(type), "Length bounds unknown for type %s", type);
-  }
-
-  private static <T> LengthBounds calculateLengthBounds(final Class<T> type) {
-    final ReflectionInformation reflectionInfo = getRequiredReflectionInfo(type);
-    LengthBounds lengthBounds = LengthBounds.ZERO;
-    int variableFieldCount = 0;
-    int vectorCount = 0;
-    int bitvectorCount = 0;
-    for (Field field : reflectionInfo.getFields()) {
-      final Class<?> fieldType = field.getType();
-      final LengthBounds fieldLengthBounds;
-      if (getOptionalReflectionInfo(fieldType).isPresent()) {
-        fieldLengthBounds = calculateLengthBounds(fieldType);
-
-      } else if (fieldType == Bitlist.class) {
-        fieldLengthBounds = calculateBitlistLength(reflectionInfo, variableFieldCount);
-
-      } else if (fieldType == SSZList.class) {
-        fieldLengthBounds = calculateSszListLength(reflectionInfo, variableFieldCount);
-
-      } else if (fieldType == SSZVector.class) {
-        fieldLengthBounds = calculateSszVectorLength(reflectionInfo, vectorCount);
-        vectorCount++;
-
-      } else if (fieldType == Bitvector.class) {
-        fieldLengthBounds = calculateBitvectorLength(reflectionInfo, bitvectorCount);
-        bitvectorCount++;
-
-      } else if (isPrimitive(fieldType)) {
-        final int fixedLength = getPrimitiveLength(fieldType);
-        fieldLengthBounds = new LengthBounds(fixedLength, fixedLength);
-
-      } else {
-        throw lengthNotImplemented(fieldType);
-      }
-
-      if (isVariable(fieldType)) {
-        variableFieldCount++;
-        // The fixed parts includes an offset in place of the variable length value
-        lengthBounds =
-            lengthBounds.add(
-                BYTES_PER_LENGTH_OFFSET.intValue(), BYTES_PER_LENGTH_OFFSET.intValue());
-      }
-      lengthBounds = lengthBounds.add(fieldLengthBounds);
-    }
-    return lengthBounds;
-  }
-
-  private static LengthBounds calculateBitvectorLength(
-      final ReflectionInformation reflectionInfo, final int bitvectorCount) {
-    final LengthBounds fieldLengthBounds;
-    final Integer size = reflectionInfo.getBitvectorSizes().get(bitvectorCount);
-    final int serializationLength = Bitvector.sszSerializationLength(size);
-    fieldLengthBounds = new LengthBounds(serializationLength, serializationLength);
-    return fieldLengthBounds;
-  }
-
-  private static LengthBounds calculateSszVectorLength(
-      final ReflectionInformation reflectionInfo, final int vectorCount) {
-    final LengthBounds fieldLengthBounds;
-    final Class elementType = reflectionInfo.getVectorElementTypes().get(vectorCount);
-    final int vectorLength = reflectionInfo.getVectorLengths().get(vectorCount);
-    final LengthBounds elementLengthBounds = getElementLengthBounds(elementType);
-    fieldLengthBounds =
-        new LengthBounds(
-            vectorLength * elementLengthBounds.getMin(),
-            vectorLength * elementLengthBounds.getMax());
-    return fieldLengthBounds;
-  }
-
-  private static LengthBounds calculateSszListLength(
-      final ReflectionInformation reflectionInfo, final int variableFieldCount) {
-    final LengthBounds fieldLengthBounds;
-    final Class listElementType = reflectionInfo.getListElementTypes().get(variableFieldCount);
-    final long listElementMaxSize = reflectionInfo.getListElementMaxSizes().get(variableFieldCount);
-    final LengthBounds elementLengthBounds = getElementLengthBounds(listElementType);
-    final long variableFieldOffsetsLength =
-        isVariable(listElementType) ? BYTES_PER_LENGTH_OFFSET.intValue() * listElementMaxSize : 0;
-    fieldLengthBounds =
-        new LengthBounds(
-            0, elementLengthBounds.getMax() * listElementMaxSize + variableFieldOffsetsLength);
-    return fieldLengthBounds;
-  }
-
-  private static LengthBounds calculateBitlistLength(
-      final ReflectionInformation reflectionInfo, final int variableFieldCount) {
-    final LengthBounds fieldLengthBounds;
-    final long maxSize = reflectionInfo.getBitlistElementMaxSizes().get(variableFieldCount);
-    fieldLengthBounds =
-        new LengthBounds(
-            Bitlist.getSerializedSize(Math.toIntExact(0)),
-            Bitlist.getSerializedSize(Math.toIntExact(maxSize)));
-    return fieldLengthBounds;
-  }
-
-  private static LengthBounds getElementLengthBounds(final Class<?> listElementType) {
-    if (isPrimitive(listElementType)) {
-      final int primitiveLength = getPrimitiveLength(listElementType);
-      return new LengthBounds(primitiveLength, primitiveLength);
-    }
-    return calculateLengthBounds(listElementType);
-  }
-
-  private static RuntimeException lengthNotImplemented(final Class<?> fieldType) {
-    return new IllegalArgumentException(
-        "Don't know how to calculate length for " + fieldType.getSimpleName());
   }
 
   private static void assertAllDataRead(SSZReader reader) {
@@ -580,7 +474,7 @@ public class SimpleOffsetSerializer {
     return SSZVector.createMutable(newList, classInfo);
   }
 
-  private static ReflectionInformation getRequiredReflectionInfo(Class classInfo) {
+  static ReflectionInformation getRequiredReflectionInfo(Class classInfo) {
     final ReflectionInformation reflectionInfo = classReflectionInfo.get(classInfo);
     checkArgument(
         reflectionInfo != null,
@@ -588,7 +482,7 @@ public class SimpleOffsetSerializer {
     return reflectionInfo;
   }
 
-  private static Optional<ReflectionInformation> getOptionalReflectionInfo(Class classInfo) {
+  static Optional<ReflectionInformation> getOptionalReflectionInfo(Class classInfo) {
     return Optional.ofNullable(classReflectionInfo.get(classInfo));
   }
 
@@ -620,33 +514,12 @@ public class SimpleOffsetSerializer {
     }
   }
 
-  private static int getPrimitiveLength(Class classInfo) {
-    switch (classInfo.getSimpleName()) {
-      case "UnsignedLong":
-        return UNSIGNED_LONG_SIZE;
-      case "ArrayWrappingBytes32":
-      case "Bytes32":
-        return Bytes32.SIZE;
-      case "Bytes4":
-        return Bytes4.SIZE;
-      case "BLSSignature":
-        return BLSSignature.BLS_SIGNATURE_SIZE;
-      case "BLSPublicKey":
-        return BLSPublicKey.BLS_PUBKEY_SIZE;
-      case "Boolean":
-      case "boolean":
-        return BOOLEAN_SIZE;
-      default:
-        throw new IllegalArgumentException("Unable to deserialize " + classInfo.getSimpleName());
-    }
-  }
-
   private static int readOffset(SSZReader reader, MutableInt bytesPointer) {
     bytesPointer.add(4);
     return reader.readInt32();
   }
 
-  private static boolean isVariable(Class classInfo) {
+  static boolean isVariable(Class classInfo) {
     if (classInfo == SSZList.class || classInfo == Bitlist.class) {
       return true;
     } else {
@@ -656,18 +529,18 @@ public class SimpleOffsetSerializer {
     }
   }
 
-  private static boolean isPrimitive(Class classInfo) {
+  static boolean isPrimitive(Class classInfo) {
     return !(SSZContainer.class.isAssignableFrom(classInfo)
         || classInfo == SSZVector.class
         || classInfo == Bitvector.class
         || classInfo == VoteTracker.class);
   }
 
-  private static boolean isVector(Class classInfo) {
+  static boolean isVector(Class classInfo) {
     return classInfo == SSZVector.class;
   }
 
-  private static boolean isBitvector(Class classInfo) {
+  static boolean isBitvector(Class classInfo) {
     return classInfo == Bitvector.class;
   }
